Extend acceptance spec & introduce endpoint
The acceptance spec now enters through a (simulated) call to an
endpoint on the whole app, rather than executing the use case
directly.

As of this commit, there is still an outstanding issue around datetime
formats - how the datetime is represented through the initial web
request, use case, gateway and finally the database needs to be
considered (and made to work!).